--- app/src/main/java/com/basaraksanli/photoAlbum/data/remote/responses/AlbumList.kt
@@ -1,3 +1,0 @@
-package com.basaraksanli.photoAlbum.data.remote.responses
-
-class AlbumList : ArrayList<AlbumListItem>()
