--- app/src/main/java/com/basaraksanli/photoAlbum/data/remote/responses/AlbumListItem.kt
@@ -1,7 +1,0 @@
-package com.basaraksanli.photoAlbum.data.remote.responses
-
-data class AlbumListItem(
-    val id: Int,
-    val title: String,
-    val userId: Int
-)
