--- app/src/main/java/com/basaraksanli/photoAlbum/data/remote/responses/Company.kt
@@ -1,7 +1,0 @@
-package com.basaraksanli.photoAlbum.data.remote.responses
-
-data class Company(
-    val bs: String,
-    val catchPhrase: String,
-    val name: String
-)
