--- app/src/main/java/com/basaraksanli/photoAlbum/data/remote/responses/Geo.kt
@@ -1,6 +1,0 @@
-package com.basaraksanli.photoAlbum.data.remote.responses
-
-data class Geo(
-    val lat: String,
-    val lng: String
-)
