--- app/src/main/java/com/basaraksanli/photoAlbum/data/remote/responses/PhotoList.kt
@@ -1,3 +1,0 @@
-package com.basaraksanli.photoAlbum.data.remote.responses
-
-class PhotoList : ArrayList<PhotoListItem>()
