--- app/src/main/java/com/basaraksanli/photoAlbum/data/remote/responses/PhotoListItem.kt
@@ -1,9 +1,0 @@
-package com.basaraksanli.photoAlbum.data.remote.responses
-
-data class PhotoListItem(
-    val albumId: Int,
-    val id: Int,
-    val thumbnailUrl: String,
-    val title: String,
-    val url: String
-)
