--- app/src/main/java/com/basaraksanli/photoAlbum/data/remote/responses/UserList.kt
@@ -1,3 +1,0 @@
-package com.basaraksanli.photoAlbum.data.remote.responses
-
-class UserList : ArrayList<UserListItem>()
