--- app/src/main/java/com/basaraksanli/photoAlbum/data/remote/responses/UserListItem.kt
@@ -1,13 +1,0 @@
-package com.basaraksanli.photoAlbum.data.remote.responses
-
-
-data class UserListItem(
-    val address: Address,
-    val company: Company,
-    val email: String,
-    val id: Int,
-    val name: String,
-    val phone: String,
-    val username: String,
-    val website: String
-)
