+++ app/src/main/java/com/basaraksanli/photoAlbum/feature_album/data/repository/AlbumRepositoryImpl.kt
@@ -1,0 +1,25 @@
+package com.basaraksanli.photoAlbum.feature_album.data.repository
+
+import com.basaraksanli.photoAlbum.feature_album.data.remote.AlbumApi
+import com.basaraksanli.photoAlbum.feature_album.domain.model.AlbumList
+import com.basaraksanli.photoAlbum.feature_album.domain.model.PhotoList
+import com.basaraksanli.photoAlbum.feature_album.domain.model.UserList
+import com.basaraksanli.photoAlbum.feature_album.domain.repository.AlbumRepository
+import dagger.hilt.android.scopes.ActivityScoped
+import javax.inject.Inject
+
+@ActivityScoped
+class AlbumRepositoryImpl @Inject constructor(
+    private val api: AlbumApi) : AlbumRepository{
+    override suspend fun gelUserList(): UserList {
+        return api.getUserList()
+    }
+
+    override suspend fun getAlbumList(): AlbumList {
+        return api.getAlbumList()
+    }
+
+    override suspend fun getPhotoList(albumId: Int): PhotoList {
+        return api.getPhotoList(albumId = albumId)
+    }
+}
