+++ app/src/main/java/com/basaraksanli/photoAlbum/feature_album/domain/model/Address.kt
@@ -1,0 +1,9 @@
+package com.basaraksanli.photoAlbum.feature_album.domain.model
+
+data class Address(
+    val city: String,
+    val geo: Geo,
+    val street: String,
+    val suite: String,
+    val zipcode: String
+)
