+++ app/src/main/java/com/basaraksanli/photoAlbum/feature_album/domain/model/AlbumListItem.kt
@@ -1,0 +1,7 @@
+package com.basaraksanli.photoAlbum.feature_album.domain.model
+
+data class AlbumListItem(
+    val id: Int,
+    val title: String,
+    val userId: Int
+)
