+++ app/src/main/java/com/basaraksanli/photoAlbum/feature_album/domain/model/Company.kt
@@ -1,0 +1,7 @@
+package com.basaraksanli.photoAlbum.feature_album.domain.model
+
+data class Company(
+    val bs: String,
+    val catchPhrase: String,
+    val name: String
+)
