+++ app/src/main/java/com/basaraksanli/photoAlbum/feature_album/domain/model/Geo.kt
@@ -1,0 +1,6 @@
+package com.basaraksanli.photoAlbum.feature_album.domain.model
+
+data class Geo(
+    val lat: String,
+    val lng: String
+)
