+++ app/src/main/java/com/basaraksanli/photoAlbum/feature_album/domain/model/PhotoListItem.kt
@@ -1,0 +1,9 @@
+package com.basaraksanli.photoAlbum.feature_album.domain.model
+
+data class PhotoListItem(
+    val albumId: Int,
+    val id: Int,
+    val thumbnailUrl: String,
+    val title: String,
+    val url: String
+)
