+++ app/src/main/java/com/basaraksanli/photoAlbum/feature_album/domain/model/UserList.kt
@@ -1,0 +1,3 @@
+package com.basaraksanli.photoAlbum.feature_album.domain.model
+
+class UserList : ArrayList<UserListItem>()
