+++ app/src/main/java/com/basaraksanli/photoAlbum/feature_album/domain/model/UserListItem.kt
@@ -1,0 +1,13 @@
+package com.basaraksanli.photoAlbum.feature_album.domain.model
+
+
+data class UserListItem(
+    val address: Address,
+    val company: Company,
+    val email: String,
+    val id: Int,
+    val name: String,
+    val phone: String,
+    val username: String,
+    val website: String
+)
