+++ app/src/main/java/com/basaraksanli/photoAlbum/feature_album/domain/use_case/AlbumUseCases.kt
@@ -1,0 +1,7 @@
+package com.basaraksanli.photoAlbum.feature_album.domain.use_case
+
+data class AlbumUseCases (
+    val getUserList : GetUserList,
+    val getAlbumList : GetAlbumList,
+    val getPhotoList : GetPhotoList
+)
