+++ app/src/main/java/com/basaraksanli/photoAlbum/feature_album/domain/use_case/GetUserList.kt
@@ -1,0 +1,18 @@
+package com.basaraksanli.photoAlbum.feature_album.domain.use_case
+
+import com.basaraksanli.photoAlbum.feature_album.domain.model.UserList
+import com.basaraksanli.photoAlbum.feature_album.domain.repository.AlbumRepository
+import com.basaraksanli.photoAlbum.feature_album.util.ApiResult
+import javax.inject.Inject
+
+class GetUserList @Inject constructor(
+    private val repository: AlbumRepository){
+    suspend operator fun invoke(): ApiResult<UserList> {
+        val response = try {
+            repository.gelUserList()
+        } catch (e: Exception) {
+            return ApiResult.Error("An unknown error occurred.")
+        }
+        return ApiResult.Success(response);
+    }
+}
