+++ app/src/main/java/com/basaraksanli/photoAlbum/feature_album/util/AppSizes.kt
@@ -1,0 +1,5 @@
+package com.basaraksanli.photoAlbum.feature_album.util
+
+object AppSizes {
+    const val USER_CARD_SIZE = 200
+}
