--- app/src/main/java/com/basaraksanli/photoAlbum/util/AppSizes.kt
@@ -1,5 +1,0 @@
-package com.basaraksanli.photoAlbum.util
-
-object AppSizes {
-    const val USER_CARD_SIZE = 200
-}
