--- app/src/main/java/com/basaraksanli/photoAlbum/util/Constants.kt
@@ -1,20 +1,0 @@
-package com.basaraksanli.photoAlbum.util
-
-import com.basaraksanli.photoAlbum.R
-
-object Constants {
-    const val BASE_URL = "https://jsonplaceholder.typicode.com/"
-    val THUMBNAIL_IMAGE_LIST : ArrayList<Int> = arrayListOf(
-        R.drawable.thumbnail1,
-        R.drawable.thumbnail2,
-        R.drawable.thumbnail3,
-        R.drawable.thumbnail4,
-        R.drawable.thumbnail5,
-        R.drawable.thumbnail6,
-        R.drawable.thumbnail7,
-        R.drawable.thumbnail8,
-        R.drawable.thumbnail9,
-        R.drawable.thumbnail10,
-    )
-
-}
